feat(app-header): migrates app header to mui
Migrates the AppHeader component to material ui components

re #143
